Move mongoose routes into /routes/mongoose
