+++ app/routes/index.js
@@ -1,46 +1,22 @@
 'use strict';
 
-const express = require('express');
-const restify = require('express-restify-mongoose');
 const receiveMessageRoute = require('./receive-message');
 const sendMessageRoute = require('./send-message');
 const importMessageRoute = require('./import-message');
 const broadcastSettingsRoute = require('./broadcast-settings');
+const mongooseRoutes = require('./mongoose');
 
 // middleware
 const authenticateMiddleware = require('../../lib/middleware/authenticate');
 const parseMetadataMiddleware = require('../../lib/middleware/metadata-parse');
-
-const router = express.Router();
-
-const Conversation = require('../models/Conversation');
-const Message = require('../models/Message');
-const Campaign = require('../models/Campaign');
-
-const resultsHeader = 'X-Gambit-Results-Count';
-const resultsLimit = 50;
-restify.serve(router, Conversation, {
-  limit: resultsLimit,
-  name: 'conversations',
-  totalCountHeader: resultsHeader,
-});
-restify.serve(router, Message, {
-  limit: resultsLimit,
-  name: 'messages',
-  totalCountHeader: resultsHeader,
-});
-restify.serve(router, Campaign, {
-  limit: resultsLimit,
-  name: 'campaigns',
-  totalCountHeader: resultsHeader,
-});
 
 module.exports = function init(app) {
   app.get('/', (req, res) => res.send('hi'));
   app.get('/favicon.ico', (req, res) => res.sendStatus(204));
 
   app.use((req, res, next) => {
-    res.header('Access-Control-Expose-Headers', resultsHeader);
+    // TODO: DRY this with the mongoose routes.
+    res.header('Access-Control-Expose-Headers', 'X-Gambit-Results-Count');
     // TODO: Eventually remove this.
     // @see https://github.com/DoSomething/gambit-conversations/issues/55
     res.header('Access-Control-Allow-Origin', '*');
@@ -49,7 +25,7 @@ module.exports = function init(app) {
     return next();
   });
   // restified routes
-  app.use(router);
+  app.use(mongooseRoutes);
   // authenticate all requests after this line
   app.use(authenticateMiddleware());
   // parse metadata like requestId and retryCount for all requests after this line
